[autorevert] fix handling for insufficient successes (#7235)
Previously the code was trying to group branches for restarts resulting
from "infra check" and from "insufficient events", and this was a
mistake, resulting in delayed restarts.

Specifically, in this situation:
<img width="999" height="747" alt="image"
src="https://github.com/user-attachments/assets/9cd0051e-8d87-4fe2-af90-88a776847c4d"
/>
a restart on the success side is expected, but the system waits for
pending job on the failure side.


This PR decouples and simplifies the logic. Now, all restarts are
scheduled independently (relying on set deduplication) and all final
checks are performed afterwards.

Added a unit test to specifically verify the case above.