[AUTOREVERT] use secret store over environment variables for passwords (#7240)
This is due to a security task (that I agree on) to move password stored
in environment variables to secret store. The 2nd have a much tight
control on access than the first one.

**IMPORTANT**

this change, to work, will require changing permissions on terraform for
lambda https://github.com/meta-pytorch/pytorch-gha-infra/pull/826