Add vLLM CI metrics (#7239)
This PR has two parts, the query `vllm/merges_percentage` and a new HUD
page `metrics/vllm.tsx` to display vLLM CI metrics. There are 2 KPIs to
start with:

1. The % of force merges with CI failures.  Its meaning is clear.
2. The % of manual merges where a vLLM maintainer merged the pull
request manually without using GitHub auto-merge. But there wasn't any
failures in the pull request at the time of the merge.

### What is a vLLM CI failure?
As vLLM CI is on Buildkite, a CI failure means a failed Buildkite job
that (1) is not a soft fail, and (2) fails on its latest retry at the
time of the merge. We can get this information by joining the GitHub
`pull_request` with Buildkite `vllm_buildkite_jobs` on the pull request
number.

### Testing

https://torchci-git-vllm-metrics-fbopensource.vercel.app/metrics/vllm

cc @rzabarazesh @yeqcharlotte @simon-mo

---------

Signed-off-by: Huy Do <huydhn@gmail.com>