[drci] Remove drci updates from some data and torchtune (#7322)
The most recent PR updates for these repos was 1 month ago. They
previously had pytorch-bot installed on them but currently do not after
the migration to meta-pytorch. I requested it to be added to torchtune,
but there have been no PR updates to this repo for more than a month, so
I'm not sure if it matters that much if drci comments get updates