[binaries] metadata lambda to delete existing metadata (#7359)
Delete previous metadata when new wheel is uploaded

This way, if the lambda fails, there will not be stale data

Testing:
checked on something that doesn't have a metadata file already
checked on something that does have metadata file already


I think it might be better if the metadata is uploaded with the binary,
maybe add some s3 metadata to the binary and whl to show theyre in sync,
then the lambda deletes metadata whose sync tag doesn't match the
binary's

If the binary is uploaded manually, then it there won't be metadata, but
I'm not sure how often this happens. Also domains that do custom stuff?
but idk how often this happens either. Also additional packages that get
uploaded to the index?