Add python-build-package variant to windows wheel building (#7256)
In #6352, building of wheels for win-arm64 was added. This PR aligns the
naming in the workflow more closely with the rest of the workflow file
and adds the more modern `python -m build` variant to the previously
used direct invocation of `setup.py bdist_wheel` as is done for other
platforms.