[autorevert] treat test failures as successes in job-track Signals (#7399)
Currently, we have a slight overlap between job and test Signals.
Specifically, when there is any non-test related failure in the job
within the lookback window (e.g. any infra flake), the job signal is
extracted, and it extracts ANY job failures, including test failures as
the job signal. This potentially leads to the duplication, when the same
test failure is processed both as test and job-track Signals, and it
increases the chance of false positives.

This PR makes the separation explicit:
1. test failures are extracted as test-track Signals
2. job-track Signals deliberately **ignore** job failures caused
**exclusively** by tests (such events are extracted as successes)

---

The intended outcome of this change:
1. job-track signals will only try to revert infra-breaking changes and
ignore test failures
2. test-track signal processing remains as it is