[old alerts] Rules for queue alerts per machine (#7397)
Not sure if this is still worth with grafana

But it is really nice that this alert prints out the machines that are
queueing, but for grafana I think you have to make 1 alert per machine
type in order for it to say? maybe there are ways with variables that
I'm not familiar with

Add rules for queue alerts based on machine type. Also change thresholds
so they're more reasonable?