[benchmarks] navbar dropdown to be slightly more generic (#7490)
Changes the navbar dropdown to be a bit more generic. It moves the
sorting out of the component and into the function that returns the
benchmark navbar metadata based on the available categories

It also simplifies the item/group/bottom logic so you can input your
exact dropdown organization

This means we can use this dropdown for other parts of the hud navbar
without worrying about the order changing due to sorting

I checked that the look of the menu is the same. The links also seem to
go to the same place